Clean the code in cross chain module.
