+++ contract/AElf.Contracts.CrossChain/CrossChainReferenceState.cs
@@ -13,7 +13,7 @@ namespace AElf.Contracts.CrossChain
         internal AEDPoSContractContainer.AEDPoSContractReferenceState ConsensusContract { get; set; }
         internal AssociationContractContainer.AssociationContractReferenceState AssociationContract { get; set; }
         internal ACS0Container.ACS0ReferenceState GenesisContract { get; set; }
-        internal ParliamentContractContainer.ParliamentContractReferenceState ParliamentContract { get; set;}
+        internal ParliamentContractContainer.ParliamentContractReferenceState ParliamentContract { get; set; }
         internal ConfigurationContainer.ConfigurationReferenceState ConfigurationContract { get; set; }
     }
 }
