+++ src/AElf.CrossChain.Communication.Grpc/Server/GrpcParentChainServerBase.cs
@@ -43,7 +43,6 @@ namespace AElf.CrossChain.Communication.Grpc
                 try
                 {
                     await responseStream.WriteAsync(parentChainBlockData);
-                    //Logger.LogTrace($"Response parent chain data {parentChainBlockData.Height}");
                     requestedHeight++;
                 }
                 catch (InvalidOperationException)
