+++ src/AElf.CrossChain.Communication/Application/ICrossChainClientService.cs
@@ -6,7 +6,7 @@ namespace AElf.CrossChain.Communication.Application
     public interface ICrossChainClientService
     {
         Task<ChainInitializationData> RequestChainInitializationData(int chainId);
-        
+
         Task RequestCrossChainDataAsync(int chainId, long targetHeight);
 
         Task CreateClientAsync(CrossChainClientDto crossChainClientDto);
