+++ src/AElf.CrossChain.Communication/CrossChainCommunicationModule.cs
@@ -1,4 +1,3 @@
-using AElf.Kernel.Node.Infrastructure;
 using AElf.Modularity;
 using Microsoft.Extensions.DependencyInjection;
 using Volo.Abp.Modularity;
