+++ src/AElf.CrossChain.Communication/Infrastructure/ICrossChainClient.cs
@@ -10,7 +10,10 @@ namespace AElf.CrossChain.Communication.Infrastructure
         int RemoteChainId { get; }
         string TargetUriString { get; }
         bool IsConnected { get; }
-        Task RequestCrossChainDataAsync(long targetHeight, Func<IBlockCacheEntity, bool> crossChainBlockDataEntityHandler);
+
+        Task RequestCrossChainDataAsync(long targetHeight,
+            Func<IBlockCacheEntity, bool> crossChainBlockDataEntityHandler);
+
         Task<ChainInitializationData> RequestChainInitializationDataAsync(int chainId);
         Task ConnectAsync();
         Task CloseAsync();
