+++ src/AElf.CrossChain.Core/CrossChainConstants.cs
@@ -6,5 +6,6 @@ namespace AElf.CrossChain
         public const int DefaultBlockCacheEntityCount = 256;
         public const int LibHeightOffsetForCrossChainIndex = 0;
         public const string CrossChainExtraDataNamePrefix = "CrossChain";
+        public const string ConsensusExtraDataNamePrefix = "Consensus";
     }
 }
